replace redigo/redis to go-redis/redis

+++ example/redis_people/gen_SortUserBlogs_struct.go
@@ -5,9 +5,9 @@ import "time"
 var _ time.Time
 
 type SortUserBlogs struct {
-	UserId int32 `bson:"UserId" json:"UserId"`
-	Score  int32 `bson:"Score" json:"Score"`
-	BlogId int32 `bson:"BlogId" json:"BlogId"`
+	UserId int32   `bson:"UserId" json:"UserId"`
+	Score  float64 `bson:"Score" json:"Score"`
+	BlogId int32   `bson:"BlogId" json:"BlogId"`
 }
 
 func (p *SortUserBlogs) GetNameSpace() string {
